Add files via upload
this folder contains the data of the one and only major project i worked on.

this project aims to create a local patent publishing entity.
this project is a small-scale project I made as a graded project for my class 12th, "Python" subject project. 

the highlight of the project is the "Password". what it does is that after creating a file you can edit it by entering a specified password.
this password can not be changed if you are trying to edit, else you may lose access to the file permanently.

this project was primarily done by me. however, i took the help of one of my class mate in creating the password section